library_management_system/library/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
Pillow==9.1.1
pytz==2021.3
sqlparse==0.4.2